requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
itsdangerous>=2.1.0
